Estruturando as pastas e preparando o basico
