+++ doaacao/src/Components/Layout/Container.js
@@ -2,7 +2,8 @@ import styles from './Container.module.css'
 
 function Container(props){
     return(
-        <div className={`${styles.Container} ${styles[props.customClass]}`}>{props.children}</div>
+        <div className={`${styles.Container} ${styles[props.customClass]}`}>
+            {props.children}</div>
     )
 }
 export default Container
