+++ doaacao/src/Components/Pages/Campanhas.js
@@ -1,0 +1,6 @@
+function Campanhas(){
+    return(
+        <h1>Campanhas</h1>
+    )
+}
+export default Campanhas
