+++ doaacao/src/Components/Pages/Doacao.js
@@ -1,0 +1,6 @@
+function Doacao(){
+    return(
+        <h1>Doacao</h1>
+    )
+}
+export default Doacao
