+++ doaacao/src/Components/Pages/Home.js
@@ -1,0 +1,6 @@
+function    Home(){
+    return(
+        <h1>Home</h1>
+    )
+}
+export default Home
