+++ doaacao/src/Components/Pages/NovaCampanha.js
@@ -1,0 +1,6 @@
+function NovaCampanha(){
+    return(
+        <h1>NovaCampanha</h1>
+    )
+}
+export default NovaCampanha
