+++ doaacao/src/Components/Pages/SignIn.js
@@ -1,0 +1,6 @@
+function SiginIn(){
+    return(
+        <h1>SiginIn</h1>
+    )
+}
+export default SiginIn
