+++ doaacao/src/Components/Pages/SignUp.js
@@ -1,0 +1,6 @@
+function SignUp(){
+    return(
+        <h1>SignUp</h1>
+    )
+}
+export default SignUp
