--- doaacao/src/Pages/Doacao.js
@@ -1,10 +1,0 @@
-function Doacao(){
-    return (
-        <div>
-            <h1>Doacao</h1>
-            <p>Conteúdo da página</p>
-        </div>
-
-    )
-}
-export default Doacao
