--- doaacao/src/Pages/Home.js
@@ -1,10 +1,0 @@
-function Home(){
-    return (
-        <div>
-            <h1>Home</h1>
-            <p>Conteúdo da página</p>
-        </div>
-
-    )
-}
-export default Home
