--- doaacao/src/Pages/SignIn.js
@@ -1,10 +1,0 @@
-function SignIn(){
-    return (
-        <div>
-            <h1>SignIn</h1>
-            <p>Conteúdo da página</p>
-        </div>
-
-    )
-}
-export default SignIn
