--- doaacao/src/Pages/SignUp.js
@@ -1,10 +1,0 @@
-function SignUp(){
-    return (
-        <div>
-            <h1>SignUp</h1>
-            <p>Conteúdo da página</p>
-        </div>
-
-    )
-}
-export default SignUp
